Relocate linking to table-C to Rev5 update layer
Instead of putting this in secondary (more generic) SP 800-53 stylesheets for HTML/FO
that are not specific to the FISMA team, we are presuming Table C control type callout
table exist its when it doesn't in that layer. It exists in the FISMA tertiary layer,
a subsequent stylesheet called with its template. Relocate it accordingly.